requirements.txt
streamlit==1.28.*
pandas
matplotlib
seaborn
plotly.express
matplotlib
wordcloud